FE: major rewrite.
After login and sign up, we simply refresh the browser window to
re-initialize the app (socket with auth token, client etc).  But
before refreshing, we ensure apollo has flushed to local storage
(this takes a little while and was the reason we were unable to query
for user immediately after login/sign up (because data was not yet
in local storage) ).