E2e cache issue solved?
I did not realize that Cypress.env(name, value) is global to all tests
in the same file so unsetting it after each test was futile. What I
settled on was simply mutating the cache before each test runs.
So we have only one cache shared by all tests in a spec file. I know
this is a really terrible idea, but this is the best I can do for now,
and it seems to have solve the problem.